refactor(engine): rename state.ts → context.ts
The file manages graph context updates — applyContextUpdates,
enforceStrictContext, buildContextSetResult, buildInspectResult.
Session state (the traversal stack) lives in traversal-store.ts; this
file is purely about context operations. The name "state.ts" was
confusing on every read.

No behavior change, single rename + import path fix in engine.ts
(with import block reordered to satisfy Biome's organizeImports).
